Update files formatting

+++ replay-api/src/main/java/me/lagggpixel/replay/api/replay/content/IReplaySession.java
@@ -19,7 +19,7 @@ public interface IReplaySession {
 
     List<BukkitRunnable> startedTasks();
 
-    int getCurrentTick();
+    long getCurrentTick();
 
     void start();
     void pause();
